Make Result type 64-bit

+++ include/common/types.hpp
@@ -25,7 +25,7 @@ using u16 = std::uint16_t;
 using u32 = std::uint32_t;
 using u64 = std::uint64_t;
 
-using Result = u32;
+using Result = u64;
 
 using i8  = std::int8_t;
 using i16 = std::int16_t;
